+++ src/App/Adapters/WebAppAdapter.php
@@ -80,8 +80,6 @@ class WebAppAdapter extends AppAdapter
             return ExitCode::SUCCESS;
         }
 
-        $this->loadLanguage();
-
         $this->logDebugInfo();
 
         $viewCache = $this->setupViewCache();